requirements.txt
openai==1.97.1
python-dotenv==1.1.1
pandas==2.3.1
numpy==2.3.2
scipy==1.16.1
matplotlib==3.10.3
seaborn==0.13.2
scikit-learn==1.7.1
statsmodels==0.14.5
langchain-openai==0.3.28
kaggle==1.7.4.5
Flask==3.1.1
flask-cors==6.0.1
## The following requirements were added by pip freeze:
annotated-types==0.7.0
anyio==4.9.0
bleach==6.2.0
blinker==1.9.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
contourpy==1.3.3
cycler==0.12.1
distro==1.9.0
fonttools==4.59.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.8
langchain-core==0.3.72
langsmith==0.4.8
MarkupSafe==3.0.2
orjson==3.11.1
packaging==25.0
patsy==1.0.1
pillow==11.3.0
protobuf==6.31.1
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
tenacity==9.1.2
text-unidecode==1.3
threadpoolctl==3.6.0
tiktoken==0.9.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
webencodings==0.5.1
Werkzeug==3.1.3
zstandard==0.23.0   